#17 [REFACTOR] 홈페이지 구조 개선
