+++ src/components/home/MediaCardCarousel.tsx
@@ -1,30 +1,20 @@
-"use client";
-
-import MediaCard from "@/components/home/MediaCard";
-import { useDragScroll } from "@/lib/hooks/useDragScroll";
+import DragScrollCarousel from "@/components/home/DragScrollCarousel";
+import MediaThumbnail from "@/components/home/MediaThumbnail";
 import { TmdbMedia } from "@/types/tmdb";
 
 interface MediaCardCarouselProps {
   items: TmdbMedia[];
+  shape?: "rect" | "circle";
   size?: "small" | "large";
 }
 
-const MediaCardCarousel = ({ items, size = "small" }: MediaCardCarouselProps) => {
-  const { ref, onMouseDown, onMouseMove, onMouseUp, onMouseLeave } = useDragScroll();
-
+const MediaCardCarousel = ({ items, shape = "rect", size = "small" }: MediaCardCarouselProps) => {
   return (
-    <div
-      ref={ref}
-      onMouseDown={onMouseDown}
-      onMouseMove={onMouseMove}
-      onMouseUp={onMouseUp}
-      onMouseLeave={onMouseLeave}
-      className="flex cursor-pointer gap-1.75 overflow-hidden select-none"
-    >
+    <DragScrollCarousel>
       {items.map(item => (
-        <MediaCard key={item.id} item={item} size={size} />
+        <MediaThumbnail key={item.id} item={item} shape={shape} size={size} />
       ))}
-    </div>
+    </DragScrollCarousel>
   );
 };
